changed errors in abstract classes
