+++ hypex/dataset/backends/abstract.py
@@ -2,6 +2,7 @@ from abc import abstractmethod, ABC
 from typing import Iterable, Any, Union, Callable, Sized, Optional, Dict, Sequence
 
 from hypex.utils import FromDictType
+from hypex.utils.errors import AbstractMethodError
 
 
 class DatasetBackendNavigation(ABC):
@@ -13,48 +14,34 @@ class DatasetBackendNavigation(ABC):
     @property
     @abstractmethod
     def index(self):
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @property
     @abstractmethod
     def columns(self):
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def from_dict(
         self, data: FromDictType, index: Optional[Union[Iterable, Sized]] = None
     ):
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def to_dict(self) -> Dict[str, Any]:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def __getitem__(self, item) -> Any:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def __len__(self) -> int:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def __repr__(self):
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def _create_empty(
@@ -62,29 +49,21 @@ class DatasetBackendNavigation(ABC):
         index: Optional[Iterable] = None,
         columns: Optional[Iterable[str]] = None,
     ):
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def _get_column_index(
         self, column_name: Union[Sequence[str], str]
     ) -> Union[int, Sequence[int]]:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def _get_column_type(self, column_name: str) -> str:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def _update_column_type(self, column_name: str, type_name: str):
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def add_column(
@@ -93,92 +72,62 @@ class DatasetBackendNavigation(ABC):
         name: str,
         index: Optional[Sequence] = None,
     ):
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def append(self, other, index: bool = False) -> Any:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def loc(self, values: Iterable) -> Iterable:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def iloc(self, values: Iterable) -> Iterable:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
 
 class DatasetBackendCalc(DatasetBackendNavigation, ABC):
     @abstractmethod
     def mean(self) -> Any:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def max(self) -> Any:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def min(self) -> Any:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def count(self) -> Any:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def sum(self) -> Any:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def agg(self, func) -> Any:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def apply(self, func: Callable, **kwargs) -> Any:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def map(self, func: Callable, **kwargs) -> Any:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def unique(self) -> Any:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def isin(self, values: Iterable) -> Any:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
 
     @abstractmethod
     def groupby(self, by: Union[str, Iterable[str]], **kwargs) -> Any:
-        raise NotImplementedError(
-            "This method is abstract and will be overridden in DatasetBase class."
-        )
+        raise AbstractMethodError
