+++ hypex/utils/errors.py
@@ -34,3 +34,7 @@ class ComparisonNotSuitableFieldError(Exception):
 class NotFoundInExperimentDataError(Exception):
     def __init__(self, class_: type):
         super().__init__(f"{class_.__name__} id is not found in ExperimentData")
+
+class AbstractMethodError(Exception):
+    def __init__(self):
+        super().__init__("This method is abstract and will be overridden in derived class.")
